backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
python-dotenv
openai
pydantic[email]
